Moved associated files to their folder
